requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
httpx==0.28.1
python-dotenv==1.0.1
pytest==8.3.5

